LayoutTests/fast/regions/cssom/get-region-flow-ranges-inline-only-expected.txt

[CSSRegions] Test for Region::getRegionFlowRanges for inline elements

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Flow test
PASS range1.length is 1
PASS Array [wrapper,0,wrapper,1] is equal to [wrapper,0,wrapper,1]
PASS range2.length is 0
PASS successfullyParsed is true

TEST COMPLETE
